refactor: 重构 publish 插件
