+++ .env
@@ -9,6 +9,6 @@ GITHUB_APPS
 INPUT_CONFIG
 
 # 插件测试结果
-PLUGIN_TEST_RESULT
-PLUGIN_TEST_OUTPUT
-PLUGIN_TEST_METADATA
+# PLUGIN_TEST_RESULT
+# PLUGIN_TEST_OUTPUT
+# PLUGIN_TEST_METADATA
