+++ src/plugins/github/config.py
@@ -1,0 +1,19 @@
+from pathlib import Path
+from pydantic import BaseModel, ConfigDict
+
+
+class PublishConfig(BaseModel):
+    base: str
+    plugin_path: Path
+    bot_path: Path
+    adapter_path: Path
+    registry_repository: str = "bigorangeqwq/registry"
+
+
+class Config(BaseModel, extra="ignore"):
+    model_config = ConfigDict(coerce_numbers_to_str=True)
+
+    input_config: PublishConfig
+    github_repository: str
+    github_run_id: str
+    skip_plugin_test: bool = False
