+++ src/plugins/github/plugins/publish/templates/comment.md.jinja
