+++ src/plugins/github/plugins/publish/templates/render_data.md.jinja
