--- src/plugins/publish/models.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-
-
-class RepoInfo(BaseModel):
-    """仓库信息"""
-
-    owner: str
-    repo: str
